pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sistema-gestion-inmobiliaria"
version = "1.0.0"
description = "Sistema de Gestión Inmobiliaria con Clean Architecture"
authors = [
    {name = "Inmobiliaria Velar SAS"}
]
requires-python = ">=3.10"
dependencies = [
    "flet>=0.21.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Activar gradualmente

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
